refactor component imports and update transition style
